Make the two example crates usable as libraries.
